Move regrid utilities to ``common`` module
