+++ examples/mf6/different_ways_to_regrid_models.py
@@ -25,8 +25,8 @@ import numpy as np
 import xarray as xr
 from example_models import create_twri_simulation
 
+from imod.common.utilities.regrid import RegridderType, RegridderWeightsCache
 from imod.mf6.regrid import NodePropertyFlowRegridMethod
-from imod.mf6.utilities.regrid import RegridderType, RegridderWeightsCache
 
 # %%
 # Now we create the twri simulation itself. It yields a simulation of a flow problem, with a grid of 3 layers and 15 cells in both x and y directions.
