+++ imod/mf6/__init__.py
@@ -54,7 +54,6 @@ from imod.mf6.src import MassSourceLoading
 from imod.mf6.ssm import SourceSinkMixing
 from imod.mf6.sto import SpecificStorage, Storage, StorageCoefficient
 from imod.mf6.timedis import TimeDiscretization
-from imod.mf6.utilities.regrid import RegridderType, RegridderWeightsCache
 from imod.mf6.uzf import UnsaturatedZoneFlow
 from imod.mf6.wel import LayeredWell, Well, WellDisStructured, WellDisVertices
 from imod.mf6.write_context import WriteContext
