+++ imod/mf6/npf.py
@@ -6,16 +6,16 @@ import numpy as np
 import xarray as xr
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
+from imod.common.utilities.regrid import (
+    RegridderWeightsCache,
+    _regrid_package_data,
+)
 from imod.logging import init_log_decorator
 from imod.mf6.package import Package
 from imod.mf6.regrid.regrid_schemes import (
     NodePropertyFlowRegridMethod,
 )
 from imod.mf6.utilities.imod5_converter import fill_missing_layers
-from imod.mf6.utilities.regrid import (
-    RegridderWeightsCache,
-    _regrid_package_data,
-)
 from imod.mf6.validation import PKG_DIMS_SCHEMA
 from imod.schemata import (
     AllCoordsValueSchema,
