+++ imod/mf6/simulation.py
@@ -24,6 +24,7 @@ import imod.mf6.exchangebase
 from imod.common.interfaces.imodel import IModel
 from imod.common.interfaces.isimulation import ISimulation
 from imod.common.statusinfo import NestedStatusInfo
+from imod.common.utilities.regrid import _regrid_like
 from imod.logging import LogLevel, logger, standard_log_decorator
 from imod.mf6.gwfgwf import GWFGWF
 from imod.mf6.gwfgwt import GWFGWT
@@ -42,7 +43,6 @@ from imod.mf6.package import Package
 from imod.mf6.regrid.regrid_schemes import RegridMethodType
 from imod.mf6.ssm import SourceSinkMixing
 from imod.mf6.utilities.mask import _mask_all_models
-from imod.mf6.utilities.regrid import _regrid_like
 from imod.mf6.validation_context import ValidationContext
 from imod.mf6.write_context import WriteContext
 from imod.prepare.topsystem.default_allocation_methods import (
