+++ imod/msw/idf_mapping.py
@@ -5,7 +5,7 @@ import numpy as np
 import xarray as xr
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
-from imod.mf6.utilities.regrid import RegridderWeightsCache, _regrid_array
+from imod.common.utilities.regrid import RegridderWeightsCache, _regrid_array
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import MetaSwapPackage
 from imod.msw.regrid.regrid_schemes import IdfMappingRegridMethod
