+++ imod/msw/output_control.py
@@ -4,7 +4,7 @@ from typing import Any, Optional, TextIO
 import pandas as pd
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
-from imod.mf6.utilities.regrid import RegridderWeightsCache
+from imod.common.utilities.regrid import RegridderWeightsCache
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import MetaSwapPackage
 from imod.msw.timeutil import to_metaswap_timeformat
