+++ imod/msw/pkgbase.py
@@ -7,12 +7,12 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from imod.mf6.dis import StructuredDiscretization
-from imod.mf6.mf6_wel_adapter import Mf6Wel
-from imod.mf6.utilities.regrid import (
+from imod.common.utilities.regrid import (
     RegridderWeightsCache,
     _regrid_like,
 )
+from imod.mf6.dis import StructuredDiscretization
+from imod.mf6.mf6_wel_adapter import Mf6Wel
 from imod.msw.fixed_format import format_fixed_width
 from imod.typing import IntArray
 from imod.typing.grid import GridDataArray, GridDataset
