+++ imod/tests/test_mf6/test_mf6_regrid_simulation.py
@@ -4,9 +4,9 @@ import numpy as np
 import pytest
 
 import imod
+from imod.common.utilities.regrid import RegridderWeightsCache
 from imod.mf6 import VerticesDiscretization
 from imod.mf6.regrid.regrid_schemes import ConstantHeadRegridMethod
-from imod.mf6.utilities.regrid import RegridderWeightsCache
 from imod.tests.fixtures.mf6_modelrun_fixture import assert_simulation_can_run
 from imod.tests.fixtures.mf6_small_models_fixture import (
     grid_data_structured,
