+++ imod/tests/test_msw/test_infiltration.py
@@ -11,7 +11,7 @@ from numpy import nan
 from numpy.testing import assert_almost_equal, assert_equal
 from pytest_cases import case, parametrize_with_cases
 
-from imod.mf6.utilities.regrid import (
+from imod.common.utilities.regrid import (
     RegridderWeightsCache,
 )
 from imod.msw import Infiltration
