+++ imod/tests/test_msw/test_landuse_options.py
@@ -5,7 +5,7 @@ import numpy as np
 import xarray as xr
 from numpy.testing import assert_almost_equal, assert_equal
 
-from imod.mf6.utilities.regrid import (
+from imod.common.utilities.regrid import (
     RegridderWeightsCache,
 )
 from imod.msw import LanduseOptions
