+++ imod/tests/test_msw/test_scaling_factors.py
@@ -6,7 +6,7 @@ import xarray as xr
 from numpy import nan
 from numpy.testing import assert_almost_equal, assert_equal
 
-from imod.mf6.utilities.regrid import (
+from imod.common.utilities.regrid import (
     RegridderWeightsCache,
 )
 from imod.msw import ScalingFactors
